requirements.txt
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cryptography==35.0.0
h11==0.12.0
idna==3.3
numpy==1.21.2
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
outcome==1.1.0
Pillow==8.4.0
pycparser==2.20
pyOpenSSL==21.0.0
pytesseract==0.3.8
requests==2.26.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
webdriver-manager==3.5.0
wsproto==1.0.0
